+++ packages/sessions-idls/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @fogo/sessions-idls
 
+## 0.0.11
+
+### Patch Changes
+
+- f28e02a: Mark sponsor as mutable when revoking or closing sessions to make sure it can receive rent
+
+## 0.0.10
+
+### Patch Changes
+
+- 02d8d5c: Rev idl version to account for different arg format for executor quote
+
 ## 0.0.9
 
 ### Patch Changes